requirements.txt

Django>=3.1,<4.0
psycopg2-binary==2.8.6
django-bootstrap4
django-g-recaptcha
djangorestframework
django-cors-headers
requests
